requirements.txt
beautifulsoup4==4.11.1
discord==1.7.3
Flask==2.1.2
google_api_python_client==2.48.0
google_auth_oauthlib==0.5.1
grpcio==1.46.1
protobuf==3.20.1
requests==2.27.1
Werkzeug==2.1.2